+++ src/Player.java
@@ -9,19 +9,13 @@ abstract class Player {
      * An integer to uniquely identify each player. Since this implementation only considers the two-player version
      * of Carcassonne, the ID should either be 1 or 2.
      */
-    protected int playerID;
-
-    //protected int currentPoints;
-
-    //protected int numberOfMeeples;
-
-    protected Player(int playerID) {
+    protected byte playerID;
+    protected Player(byte playerID) {
         if (playerID != 1 && playerID != 2) {
             logger.error("Player initialised with invalid ID");
         }
         this.playerID = playerID;
-        //this.currentPoints = 0;
     }
 
-    abstract int[] decideOnNextMove(GameState state, GameStateSpace stateSpace, Tile tile) throws Exception;
+    abstract byte[] decideOnNextMove(GameState state, GameStateSpace stateSpace, Tile tile) throws Exception;
 }
